+++ packages/auth/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @nostr-wot/auth
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @nostr-wot/signers@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes